+++ karrio/modules/connectors/dpd/karrio/providers/dpd/shipment/__init__.py
@@ -1,0 +1,4 @@
+from karrio.providers.dpd.shipment.create import (
+    parse_shipment_response,
+    shipment_request,
+)