+++ karrio/modules/connectors/dpd/tests/__init__.py
@@ -1,0 +1,4 @@
+from tests.dpd.test_login import *
+from tests.dpd.test_rating import *
+from tests.dpd.test_tracking import *
+from tests.dpd.test_shipment import *