+++ karrio/packages/types/graphql/types.ts
@@ -5206,9 +5206,6 @@ export enum WeightUnitEnum {
 }
 
 export enum CarrierNameEnum {
-  allied_express = "allied_express",
-  allied_express_local = "allied_express_local",
-  amazon_shipping = "amazon_shipping",
   aramex = "aramex",
   asendia_us = "asendia_us",
   australiapost = "australiapost",
@@ -5224,6 +5221,7 @@ export enum CarrierNameEnum {
   dhl_universal = "dhl_universal",
   dicom = "dicom",
   dpd = "dpd",
+  dtdc = "dtdc",
   easypost = "easypost",
   easyship = "easyship",
   eshipper = "eshipper",
@@ -5232,6 +5230,7 @@ export enum CarrierNameEnum {
   generic = "generic",
   geodis = "geodis",
   hay_post = "hay_post",
+  landmark = "landmark",
   laposte = "laposte",
   locate2u = "locate2u",
   mydhl = "mydhl",